+++ server/ReusableLearningAppServer/accounts/views.py
@@ -1,0 +1,9 @@
+from rest_framework import viewsets, permissions
+from django.contrib.auth.models import User
+from accounts.serializers import UserSerializer
+
+
+class UserViewSet(viewsets.ModelViewSet):
+    queryset = User.objects.all().order_by('-date_joined')
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
