+++ server/ReusableLearningAppServer/courses/serializers.py
@@ -1,13 +1,6 @@
-from django.contrib.auth.models import User
 from rest_framework import serializers
 
 from courses.models import Course
-
-
-class UserSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ['url', 'username', 'email']
 
 
 class CourseSerializer(serializers.HyperlinkedModelSerializer):
