+++ server/ReusableLearningAppServer/courses/views.py
@@ -1,13 +1,6 @@
-from django.contrib.auth.models import User
 from rest_framework import viewsets, permissions
-from courses.serializers import UserSerializer, CourseSerializer
+from courses.serializers import CourseSerializer
 from courses.models import Course
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all().order_by('-date_joined')
-    serializer_class = UserSerializer
-    permission_classes = [permissions.IsAuthenticated]
 
 
 class CourseViewSet(viewsets.ModelViewSet):
